Refactor theme CLI switches to use chromium_src override pattern
Addresses PR review feedback by following Brave's preferred chromium_src
approach instead of direct patches:

- Replace chrome_switches patches with chromium_src shadow file using
  inline constexpr declarations
- Move ThemeCommandLineHandler implementation into anonymous namespace
  within chromium_src/chrome/browser/chrome_browser_main.cc
- Use macro-based injection points instead of direct code patches
- Remove separate .h/.cc files and BUILD.gn patches
- Minimize patch footprint to just 2 simple macro placeholders

This maintains all theme CLI functionality while following Brave's
standard override patterns for better maintainability.

Features still supported:
- --set-user-color=r,g,b: Set custom theme color
- --set-color-scheme=[system|light|dark]: Set color scheme
- --set-default-theme: Reset to default theme
- --set-grayscale-theme: Enable grayscale theme
- --set-color-variant=[tonal_spot|neutral|vibrant|expressive]: Set color variant
