--- patches/chrome-common-chrome_switches.cc.patch
@@ -1,30 +1,0 @@
-diff --git a/chrome/common/chrome_switches.cc b/chrome/common/chrome_switches.cc
-index ceb194761bcf5..cb204355064d5 100644
---- a/chrome/common/chrome_switches.cc
-+++ b/chrome/common/chrome_switches.cc
-@@ -376,6 +376,25 @@ const char kInitIsolateAsForeground[] = "init-isolate-as-foreground";
- // The format is "r,g,b", where r, g, b are a numeric values from 0 to 255.
- const char kInstallAutogeneratedTheme[] = "install-autogenerated-theme";
- 
-+// Sets the browser color scheme for all running Chrome windows.
-+// Valid values are "system", "light", or "dark".
-+const char kSetColorScheme[] = "set-color-scheme";
-+
-+// Sets the browser color variant preference.
-+// Valid values are "tonal_spot", "neutral", "vibrant", "expressive".
-+const char kSetColorVariant[] = "set-color-variant";
-+
-+// Resets to the default theme for all running Chrome windows.
-+const char kSetDefaultTheme[] = "set-default-theme";
-+
-+// Enables or disables grayscale theme for all running Chrome windows.
-+// Valid values are "true" or "false".
-+const char kSetGrayscaleTheme[] = "set-grayscale-theme";
-+
-+// Sets the user color for Chrome Refresh theming.
-+// The format is "r,g,b", where r, g, b are numeric values from 0 to 255.
-+const char kSetUserColor[] = "set-user-color";
-+
- // Causes Chrome to initiate an installation flow for the given app.
- const char kInstallChromeApp[] = "install-chrome-app";
-
