--- patches/chrome-common-chrome_switches.h.patch
@@ -1,16 +1,0 @@
-diff --git a/chrome/common/chrome_switches.h b/chrome/common/chrome_switches.h
-index 1afb66b97ccb9..6c9e45ad81878 100644
---- a/chrome/common/chrome_switches.h
-+++ b/chrome/common/chrome_switches.h
-@@ -124,6 +124,11 @@ extern const char kHomePage[];
- extern const char kIncognito[];
- extern const char kInitIsolateAsForeground[];
- extern const char kInstallAutogeneratedTheme[];
-+extern const char kSetColorScheme[];
-+extern const char kSetColorVariant[];
-+extern const char kSetDefaultTheme[];
-+extern const char kSetGrayscaleTheme[];
-+extern const char kSetUserColor[];
- extern const char kInstallChromeApp[];
- extern const char kInstallIsolatedWebAppFromFile[];
- extern const char kInstallIsolatedWebAppFromUrl[];
